+++ src/components/EmptyState.tsx
@@ -1,28 +1,36 @@
 import React from "react";
-import { Box, Center } from "@chakra-ui/react";
+import { Box, Center, Flex, FlexProps } from "@chakra-ui/react";
 import colourTheme from "@styles/colours";
 
-type EmptyStateProps = {
+type EmptyStateProps = FlexProps & {
     styleProps?: Record<string, unknown>;
 };
 
 export const EmptyState: React.FC<EmptyStateProps> = ({
     children,
+    ...props
 }): JSX.Element => {
     return (
-        <Box
+        <Flex
             bg={colourTheme.colors.LightGray}
             w="100%"
             py={"40px"}
             px={"64px"}
             mt={5}
+            display="flex"
+            alignItems="center"
+            justifyContent="center"
+            textAlign="center"
             color={colourTheme.colors.Gray}
+            {...props}
         >
             <Box>
                 <Center whiteSpace="pre-line" textAlign="center">
                     {children}
                 </Center>
             </Box>
-        </Box>
+        </Flex>
     );
 };
+
+export default EmptyState;
